pyproject.toml
[tool.poetry]
name = "kucoin-futures-lib"
version = "0.12.0"
description = "Kucoin Futures wrapper library"
authors = ["Evgeny Aleshin"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
kucoin-futures-python = "^1.0.9"
python-dotenv = "^1.0.1"


[tool.poetry.group.develop.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]  # Ignore DeprecationWarnings until kucoin-futures-python is updated